split cmd/ into a separate module
Ignite want to reuse some of the library packages in this module,
such as the engine package.
However, they can't require the overall module,
as cmd/provider pulls heavy dependencies like go-ipfs.

Split the module into two; one for the actual daemon we can run,
and another for the set of library packagse others may reuse.

Note this also required moving some library packages out of "internal",
as otherwise cmd/provider can't reasonably import internal/foo.

Also note that we introduce a dependency bump dance;
whenever cmd needs to update the root library's version,
we need to first push one version of the library to git,
and then "go get" that pushed version into cmd.
For now, we've done that via the module-split branch.

Once Go 1.18 releases, we'll be able to add a root go.work file to
simplify the development of the two modules at the same time.
Note that we'll still need to do the "git push" and "go get" dance
if we want cmd/provider to be useful as a standalone module.

While here, rename "package suppliers" to "supplier",
as per Andrew's suggestion.
