+++ e2e_retrieve_test.go
@@ -7,15 +7,15 @@ import (
 	"time"
 
 	datatransfer "github.com/filecoin-project/go-data-transfer"
-	provider "github.com/filecoin-project/indexer-reference-provider"
-	"github.com/filecoin-project/indexer-reference-provider/config"
-	"github.com/filecoin-project/indexer-reference-provider/engine"
-	"github.com/filecoin-project/indexer-reference-provider/internal/cardatatransfer"
-	"github.com/filecoin-project/indexer-reference-provider/internal/libp2pserver"
-	"github.com/filecoin-project/indexer-reference-provider/internal/suppliers"
-	"github.com/filecoin-project/indexer-reference-provider/metadata"
-	p2pserver "github.com/filecoin-project/indexer-reference-provider/server/provider/libp2p"
-	"github.com/filecoin-project/indexer-reference-provider/testutil"
+	provider "github.com/filecoin-project/index-provider"
+	"github.com/filecoin-project/index-provider/cardatatransfer"
+	"github.com/filecoin-project/index-provider/config"
+	"github.com/filecoin-project/index-provider/engine"
+	"github.com/filecoin-project/index-provider/libp2pserver"
+	"github.com/filecoin-project/index-provider/metadata"
+	p2pserver "github.com/filecoin-project/index-provider/server/provider/libp2p"
+	"github.com/filecoin-project/index-provider/supplier"
+	"github.com/filecoin-project/index-provider/testutil"
 	stiapi "github.com/filecoin-project/storetheindex/api/v0"
 	p2pclient "github.com/filecoin-project/storetheindex/providerclient/libp2p"
 	"github.com/ipfs/go-datastore"
@@ -33,7 +33,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func setupServer(ctx context.Context, t *testing.T) (*libp2pserver.Server, host.Host, *suppliers.CarSupplier, *engine.Engine) {
+func setupServer(ctx context.Context, t *testing.T) (*libp2pserver.Server, host.Host, *supplier.CarSupplier, *engine.Engine) {
 	h, err := libp2p.New(ctx, libp2p.ListenAddrStrings("/ip4/0.0.0.0/tcp/0"))
 	require.NoError(t, err)
 	priv, _, err := test.RandTestKeyPair(crypto.Ed25519, 256)
@@ -48,7 +48,7 @@ func setupServer(ctx context.Context, t *testing.T) (*libp2pserver.Server, host.
 	require.NoError(t, err)
 	err = e.Start(ctx)
 	require.NoError(t, err)
-	cs := suppliers.NewCarSupplier(e, store, car.ZeroLengthSectionAsEOF(false))
+	cs := supplier.NewCarSupplier(e, store, car.ZeroLengthSectionAsEOF(false))
 	err = cardatatransfer.StartCarDataTransfer(dt, cs)
 	require.NoError(t, err)
 	s := p2pserver.New(ctx, h, e)
