+++ p2putil/message_writer.go
