+++ server/admin/http/server.go
@@ -6,9 +6,9 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/filecoin-project/indexer-reference-provider/config"
-	"github.com/filecoin-project/indexer-reference-provider/engine"
-	"github.com/filecoin-project/indexer-reference-provider/internal/suppliers"
+	"github.com/filecoin-project/index-provider/config"
+	"github.com/filecoin-project/index-provider/engine"
+	"github.com/filecoin-project/index-provider/supplier"
 	"github.com/gorilla/mux"
 	logging "github.com/ipfs/go-log/v2"
 	"github.com/libp2p/go-libp2p-core/host"
@@ -23,7 +23,7 @@ type Server struct {
 	e      *engine.Engine
 }
 
-func New(cfg config.AdminServer, h host.Host, e *engine.Engine, cs *suppliers.CarSupplier) (*Server, error) {
+func New(cfg config.AdminServer, h host.Host, e *engine.Engine, cs *supplier.CarSupplier) (*Server, error) {
 	listen, err := cfg.ListenNetAddr()
 	if err != nil {
 		return nil, err
