+++ server/provider/libp2p/protocol_test.go
@@ -4,12 +4,12 @@ import (
 	"context"
 	"testing"
 
-	"github.com/filecoin-project/indexer-reference-provider/config"
-	"github.com/filecoin-project/indexer-reference-provider/engine"
-	"github.com/filecoin-project/indexer-reference-provider/internal/libp2pserver"
-	"github.com/filecoin-project/indexer-reference-provider/internal/utils"
-	p2pserver "github.com/filecoin-project/indexer-reference-provider/server/provider/libp2p"
-	"github.com/filecoin-project/indexer-reference-provider/testutil"
+	"github.com/filecoin-project/index-provider/config"
+	"github.com/filecoin-project/index-provider/engine"
+	"github.com/filecoin-project/index-provider/libp2pserver"
+	p2pserver "github.com/filecoin-project/index-provider/server/provider/libp2p"
+	"github.com/filecoin-project/index-provider/testutil"
+	"github.com/filecoin-project/index-provider/utils"
 	stiapi "github.com/filecoin-project/storetheindex/api/v0"
 	p2pclient "github.com/filecoin-project/storetheindex/providerclient/libp2p"
 	"github.com/ipfs/go-datastore"
