+++ server/provider/libp2p/server.go
@@ -3,8 +3,8 @@ package p2pserver
 import (
 	"context"
 
-	"github.com/filecoin-project/indexer-reference-provider/engine"
-	"github.com/filecoin-project/indexer-reference-provider/internal/libp2pserver"
+	"github.com/filecoin-project/index-provider/engine"
+	"github.com/filecoin-project/index-provider/libp2pserver"
 	"github.com/libp2p/go-libp2p-core/host"
 )
